Assorted hacking and refactoring.
Filters are now broken out into an independent dll, and the whole indexing
process is a bit more rational.  The OpenOffice filter now properly flags
bold/underline/header text as hot.  The IndexDriver now tries to do the right
thing if the Hits coming back from a query appear to be out-of-date.

Lots of tweaks as I discard_my_old_ways and learn NewWaysToNameThings.