In UriFu.StringToUris, use UriStringToUri to convert each piece of the split string into a Uri.  This fixes another bug with filenames containing @.
In FileAttributesStore.AttachTimestamp, explicitly set the path on the
object we get out of ReadOrCreate.  Otherwise an already-indexed file that
is copied to another name will never appear to be up-to-date.

In QueryResult:
(1) Be more paranoid about catching exceptions
(2) Added a comment explaining why we can get away w/ not using
an ExceptionHandlingThread.