Handle inotify queue overflow events in the file system queryable.
A bunch of logging fixes.  The IndexHelper now logs to a separate file.
(Fixes bug 168098)

Proper handling of --fg and --bg in the beagled script.
(Fixes bug 168282)

Watch VmRSS to decide when to restart the IndexHelper, not VmSize.

Set the shutdown memory threshold relative to the initial memory size,
not a fixed size.

Never shutdown the helper process if it hasn't actually done anything
yet.