Changed FilterOpenOffice to be more careful about popping or peeking the hot_nodes stack.
Reduced the no-index threshold for text/plain files to 5mb.

Show a snippet in the presentation tile.

Don't show any debug spew if we can't store a file's attributes
in EAs.

In LuceneDriver, dump the term stats in a slightly more convenient order.

Make Filters more robust against exceptions thrown while pulling text.

When indexing mail, skip any part whose content type is not text.
This vastly decreases the size of the mail index.

Fixed a stupid bug in the TextCache: of course there can be more than one
self-cached file.

In RemoteIndexerImpl, don't mark the indexer as unsafe to close until
the beginning of a flush.  This will allow the helper process to shut
down gracefully if the beagled crashes somewhere between adding items
to index and calling flush.