New mail filter, which gives us MailDir support, and port the Evolution backends over to using it.
If an mbox or summary is updated and the generator is already in the
scheduler, don't readd it and reset it back to the start.  So we'll
actually index everything while Evo is running rather than recrawling.

Tweak the scheduler to throttle when many immediate priority tasks come
in.  This will help with the situations when you untar a tarball or
rm -r a directory.

Update ExtractContent to follow child indexables

Add a Uri field to the Filter, so that child indexables can set their
Uris based on the parent's.

Don't instantiate a IndexReader when we're just using it for
IndexSearcher.  That way we don't have to manage lifecycle.