glue/inotify-syscalls.h: add syscall info for the Alpha, also known as the
glue/inotify-syscalls.h: add syscall info for the Alpha, also known as the
Alpha AXP, is a 64-bit RISC microprocessor originally developed and
fabricated by Digital Equipment Corp, which used it in its own line of
workstations and servers. Designed as a successor to the VAX line of
computers, it supported the VMS operating system, as well as Digital UNIX
and, today, Linux.