glue/inotify-syscalls.h: add inotify system call support for the SH
glue/inotify-syscalls.h: add inotify system call support for the SH
architecture, also known as the SuperH.  The SuperH core is a RISC-based
architecture found in a large number of embedded systems.  The SuperH family
was first developed by Hitachi as the successor to the H8 Family and was
outsourced to the newly-formed SuperH Inc., owned by Hitachi and ST
Microelectronics. SuperH Inc now sells the designs of the CPU cores.

The lineage is as follows

SH-1 - 32-bit with maximum of 20MHz (As used on Sega Saturn to control the
CD-drive and to check the Copy Protection on the game's CD).

SH-2 - 32-bit with up to 28.7MHz (As used in the Sega Saturn).

SH-3 - 32-bit with up to 200MHz. This spring introduced a MMU to the SH
Family (As used in many Windows CE devices).

SH-4 - 32-bit dual-issue core with a 128-bit vector FPU (As used in the
Dreamcast and on some Sega Arcade Machines such as the Naomi and Naomi 2).