Scheduler.cs: Add a "Maintenance" priority to the scheduler, and track the
Scheduler.cs:
Add a "Maintenance" priority to the scheduler, and track the source of every
task.  A maintenance task is only executed after all immediate or delayed
tasks *for that source* have been executed.  Thus maintenance tasks will only
execute after all crawlers, etc. are finished.  In most backends this will
only happen when the data is fully indexed/crawled and the index is in a
complete known state.

LuceneQueryable.cs:
Attach a source to the tasks we create.
Move index optimizes into a maintenance task, to avoid unnecessary optimizes
during indexing.