Add --enable-xml-dump configure option for seeing the XML messages before they are sent.
Change the XML format to name subparts as Part rather than SubPart
in OR parts to make things easier in C.

Add OR parts to libbeagle, and generally clean up the query part
stuff there a lot.  Add docs.

Fix up the PowerPoint filter a little bit to correctly report
error when it can't parse