Fix FileSystem.GetLastWriteTimeUtc() to throw an ArgumentNullException if the path passed in is null, and a FileNotFoundException (instead of IOException) if the file isn't there.
Fix a nasty bug in the LuceneFileQueryable: if the parent URI is not set
or is set to a non-file URI, PreAddIndexableHook() was returning false.
This meant that indexables with no ParentUri or a non-file ParentUri
weren't being indexed, including *all* child indexables.

Fix a bug with the file_info_cache in the LFQ.  It was effectively
useless for shared indexables and it leaked in all cases until files
were removed.  (Which they rarely are for LFQ.)

Fix a bug in the AddGeneratorTask in which indexables created from an
indexable generator weren't cleaned up after if PreAddIndexableHook()
returned false.

Change the KMailQueryable back to a LFQ.  Thanks Bera for getting on my
case about this.