Add infrastructure to signal clients whether or not we're doing the initial index:
* Add a new request message, InformationalMessagesRequest, which is an
asynchronous message type for getting out-of-band, general messages from
the daemon.

* Add a new response message, IndexingStatusResponse, which indicates
whether or not the daemon is doing the initial index.

* Remove the use of QueryableState throughout the code; it never worked
well and so it's been deprecated.

* Make it the responsibility of the backend to report its crawl status,
rather than trying to guess it using QueryableState.  This makes things
a lot more reliable as backends can track this better than the
infrastructure.  Add support to determine this to the files, evolution
mail, evolution data server, and gaim log backends.

* Make progress percentages pulled on-demand from the backend rather than
stored as part of the LuceneQueryable.  This feels like a cleaner design.

* Hook the beagle-search UI up to the new messages, so that it can
detect changes and present the UI to the user.

And also:

* Clean up some usages of the Log class in the various backends, since I
had to touch most of them up anyway, and in the Inotify code.

* Improve some logging in the code, particularly in the FSQ backend.