Add a new archived file tile to beagle-search.  Still needs a little love.
Fix an issue where ParentUris of files were the internal URIs, not the
real URIs.

Fix a crash when the OpenWith menu has no items.

Mark items as being inside an archive in the archive filter.

Change the "your data is being indexed" text in the notification area to
not mention "Beagle", since it's not used anywhere else in the UI.