Deprecate the old --autostarted flag to beagled and beagle-search.
Now that the autostart mechanism is fairly widely deployed, ship
autostart .desktop files for beagled and beagle-search in the tarball,
and hook up the checkbox in beagle-settings to write out the autostart
files in ~/.config/autostart.

Remove old webservices code and widgets from beagle-settings, and fix
up a few accelerators.  Fix a problem in which beagle-settings wouldn't
run out of the source tree because of a missing reference to the images
assembly.

Fix the text in beagle-search's notification dialog to not mention
"Beagle", since it isn't mentioned anywhere else in the search app.