Cache content inside archives, so we can get snippets.
Composite the MIME icon for an archive on top of the thumbnail in the
archived file tile.  Thanks Federico for pointing out to me what was
causing a crash.

Fix similar code in the image tile which could also crash, but it's very
unlikely to do so.

Change the details pane of the archive tile to be a little bit more
relevant for archives.