Add a count to the number of child indexables found with beagle-extract-content.
Update the code in IndexerRequest a little; it was wrong in places, and
make it a little more judicious in its memory usage.

Fix a mem leak of URIs in LuceneFileQueryable.

Make LuceneIndexingDriver be a little smarter about getting indexables
from the IndexerRequest because of my changes to that class.

Don't store the list of accounts in the Evo mail's IMAP indexable
generator, because it's never used outside of one block of code.  Just
make it local to that function, saving a good bit of memory for every
IMAP folder.