Add support for Linux's relatively new SCHED_BATCH scheduler policy.
Add glue to set SCHED_BATCH or the default SCHED_OTHER, and hook it
into SystemPriorities.cs (which was renamed from IoPriority.cs).

In the index helper, set our nice level before we start the server,
since we want the server thread to inherit our lower priority.  In
there, try to set the scheduler policy to SCHED_BATCH.  If that
succeeds, set our nice to +12, otherwise to +17.  (The reason
for the difference is because SCHED_BATCH + 12 will equal +17.)


svn path=/trunk/; revision=3280