Move MS Word text extraction out-of-process.  libwv1 is far too unreliable to have it even inside our index helper process.  Crashes can corrupt the index and some documents cause it to explode with memory.
* Add a new beagle-doc-extractor tool, which extracts the text from
a Word document and prints it out to stdout.  Ported FilterDOC to
use it.

* Added the ability to run child setup functions with SafeProcess, since
this is a nice feature with g_spawn().  These functions run in the child
process following the fork() but before the exec().  This allows us to
do things like setrlimit(2) on child processes.

* Added glue around setrlimit(2) so we can keep out of process helpers
under a certain amount of CPU and addressable space, killing them if
they exceed it.

* Added CPU time limits to all the filters which call out of process
helpers.  The numbers are best guesses from various files I have on
my system, but will inevitably need some tweaking.

* Added a memory limit to the Word extractor, since I have a file
which causes libwv1 to eat over a gig of memory in about 3 seconds.

* Various tweaks to scripts so that beagle-doc-extractor works without
running "make install"



svn path=/trunk/; revision=3290