Fix some bugs that will hopefully kill off the dreaded looping problem. And if it doesn't, provide more information to make it easier to debug.
* LuceneQueryable.cs: Always flush the pending request queue when we
break out of the loop inside an IIndexableGenerator Add task.  Before
things would be held up until the queue was full, meaning documents
could be delayed from being indexing for quite some time.

* FileAttributesStore_Mixed.cs: If we successfully save file attributes
to xattr, drop any matching attributes from the sqlite DB.  (Found by
Brian J. Murrell)

* FileSystemQueryable/FileSystemQueryable.cs: Set the FSQ's "Debug"
attribute if the BEAGLE_DEBUG_FSQ environment variable is set.

* FileSystemQueryable/FileCrawlTask.cs: Log if it appears that we're
recrawling a directory we've seen before.  (Which should only happen if
setting up an inotify watch fails.)

* FileSystemQueryable/TreeCrawlTask.cs: Dial down the verbose logging
unless FSQ.Debug is set.

* FileSystemQueryable/DirectoryModel.cs: If FSQ.Debug is set, print out
the state that we set the directory to, implement ToString (), and fix a
bug in the DirectoryModel.CompareTo() function, which was
unconditionally preferring depth over last crawl time.  This would cause
a certain directory that was at a higher depth than other directories to
continously loop.  Change the method to only do that if the directory
state is not clean (ie, Unknown or Dirty).  Also, consider all
directories with a depth greater than 3 to be the same, since generally
the order of indexing those doesn't matter as much as higher-level
directories.

* IndexHelper/IndexHelper.cs: Print out version/runtime info like we do
for the daemon.


svn path=/trunk/; revision=3318