GetLastWriteTimeUtc() no longer throws an IOException if the file isn't there in 2.0.  It returns a silly date from 1601.  Fix various places (mostly in the file system backend) where we were depending on that behavior.
Slightly reorder the flow in FilterFactory.FilterIndexable so that we
avoid unnecessary work, and also check in here if the file has
disappeared out from underneath us.

Fix a DateTime UTC conversion problem that was still hanging around in
FileAttributesStore.

Fix our ParseExact() in StringFu.  We want to AssumeUniversal AND
AdjustToUniversal so that dates we get back are right.  Mono is broken
with these options right now -- see
http://bugzilla.ximian.com/show_bug.cgi?id=80677 -- but things work well
enough now with the broken behavior, and the AdjustToUniversal is
necessary for when it's fixed.


svn path=/trunk/; revision=3348