Put the looping bug out to pasture once and for all.
In FSQ.DoneCrawlingOneDirectory(), in the event that we get a null file
attribute back -- which should no longer happen -- complain loudly but 
still mark the directory as clean.  This will prevent the daemon from
looping on that one directory.

Fix our sqlite schema for our fallback FileAttributesStore.  We were
declaring our columns as STRING -- which isn't valid -- instead of TEXT.
This meant that fields were defaulting to being integer-like.  Filenames
like "0009" became "9" in the database.  Bump up the schema version and
fix that.  Also make our file_path index unique, so we don't have to
deal with potential multiple entries in the database.

Add debugging output to the FileAttributesStore.


svn path=/trunk/; revision=3354