Rework the way we do reporting of whether or not we're indexing to make it a more active task for backends, rather than passive.  This fixes problems trying to deal with this sanely in the file system backend.
* Remove some dead code from QueryDriver.cs.

* No longer declare IsIndexing_get as virtual, and replace it with a
private instance variable.  It's now the backends' responsibility to
set IsIndexing as the status changes, rather than waiting for the
getter to be called and reporting the current state of things.

* Update the file system backend to the new scheme.  Should fix problems
where the "currently indexing" info box stays on in beagle-search even
after the initial crawl has finished.

* Also update Evolution mail, EDS, Gaim, and Konversation backends to
the new scheme.

* Improve the status message when indexing files inside a directory.


svn path=/trunk/; revision=3377