Some performance improvements:
* When debug is on, get a lot more granular about how long something
takes.  Obviously helpful in identifying the hotspots.

* Replace our URI searches in the secondary index with a term
enumerator.  This is a lot faster and more memory friendly than doing
queries and multiple passes.

* Use a single TermDocs instance in ScoreHits(); before we were
instantiating a new one inside the loop and not closing it.  This reuses
a single one and seeks within it, closing it when we're all done.
Should use a little less memory in multiterm searches.


svn path=/trunk/; revision=3515