Largely revert LuceneQueryable back to r3026 and FileSystemQueryable back to r3027.  These were delaying the return of the receipts until after children had been indexed, but they were still being written to the index before children.  This commit fixes that.
* IndexerReceipts.cs: Add a new IndexerDeferredReceipt, which says that
a URI hasn't been indexed, but has been deferred until later.  (Ie, when
all its children, children's children, etc. have been indexed.)

* LuceneQueryable.cs: Largely revert to r3026, which simplifies the
Flush() method tremendously.  Handle the new deferred receipts, and
break the assumption that add receipts always come back from the request
that sent them.

* LuceneIndexingDriver.cs: First, clean up the namespace imports, they
were out of control.  Refactor the Flush() method quite a bit.  If child
indexables are created by the filter, return a deferred receipt and add
it to a queue of deferred indexables to be added rather than indexing it
and returning an added receipt.  Move the actual indexing of documents
out into a separate method.  After walking all of the indexables in the
request, walk over the items in the deferred queue that are ready to be
added.

* FileSystemQueryable/FileSystemQueryable.cs: Reverted some parts to
r3027, to move most of the PostChildrenIndexedHook() code back into
PostAddHook()


svn path=/trunk/; revision=3553