NautilusMetadataQueryable.cs: Don't do external->internal URI mapping in this backend; leave it up to the FSQ.
FileSystemQueryable.cs: Refactor PreAddIndexableHook() some.  Don't
create file attributes for files that haven't been crawled yet, it
doesn't work out the way we want it to.  Instead, set aside indexables
which have external URIs that don't (yet) map to internal URIs, and
process them when we do eventually index those files.  Make the
ExternalToInternalUri() method private.

Server.cs: Don't log an exception if the reason why we couldn't send a
response is because the client side closed the connection.  Just log a
simple message and continue.

Indexable.cs: Inside Merge(), only set the Timestamp to the one being
merged in if it's newer.


svn path=/trunk/; revision=3584