Lots of tweaks to the scheduler.  Do a better job of throttling when the scheduler is given a lot of immediate-scheduled tasks, especially considering that most tasks now are indexable generator tasks.
In the file system backend, create a new RemovalGenerator, which is an
IndexableGenerator for file/directory removals triggered by inotify.
This greatly reduces memory usage and scheduler overhead when doing an
"rm -r" on a directory.  Something similar will be upcoming for file
additions (which is a common problem when untarring huge tarballs).


svn path=/trunk/; revision=3656