Lots of fixes to the EDS backend.  I'm hoping this will fix the fairly mysterious crasher problems that Ubuntu users are seeing.
* Now require evolution-sharp 0.12, which implies Evolution 2.8 or 2.10.

* Use IndexableGenerators for scheduling the adds and removes.  This
makes the backend quite a bit more efficient and deals with EDS's broken
behavior with webcal calendars.

* Correctly index both memos and tasks in addition to calendar event
items.  Bump the index version because of this.

* Correctly handle removal events from the calendar backend.  Some time
ago the ABI changed from sending just the UID of the calendar event to a
structure of UID and recurrence ID (RID).  The Beagle code was never
updated for this and was using garbage for its URIs when doing these
removals.  This is a likely cause of the random crashes.

* Uncomment removals from the GetChanges call; EDS should be fixed for
this now.


svn path=/trunk/; revision=3662