Instead of using Uri to keep track of requested indexables and receipts, use a unique numeric id. This will prevent some race conditions in deferred-indexables.
PostAddHook and PostRemoveHook was meant to deal with remapping of Uri, so make them explicitly return the remapped uri.

Do not sort Indexables by their timestamp in IndexerRequest before sending them out. It is unclear why it was done.

Rest of the changes are needed to comply with the new use of Indexable.Id.

svn path=/trunk/; revision=3665