Merging the 'beagle-networking-branch' to 'trunk'.
Network services are now queried and results appear in both - beagle-query and beagle search. To be able to query remote services you have to register with them. You can do this using either beagle-config or beagle-settings. Both tools use Avahi to discover available services and provide them in fashionly manner.

To build with network services, don't forget to pass the --enable-avahi argument to configure/autogen.sh.

To make the Beagle daemon provide network services you have to launch it with the --network argument. Other than that everything is automated. But remember, you have to register it on the client side from which you are querying (as mentioned above).

I have tested the code, but as always keep your eyes open for exceptions and crashes (which are rare, but I have seen some - most likely from Avahi - but could not reproduce them).

Best of luck! :-)


svn path=/trunk/; revision=3851