Add support for Epiphany 2.22
The API has changed in the new Epiphany python bindings, so we have another
version of the plugin and we use that for the new version.

The main API change is that the "tab" object has been removed so we get a
direct handle to the EphyEmbed object with window.get_active_child().

svn path=/trunk/; revision=4646