Make beagle_quert_part_foo_set_bar() function allocate/free their own strings
Several of the beagle_query_part_foo_set_bar() functions used to take a string
alloc'ed by clients and assume that these would not be free'd by the client.
This is counter-intuitive and not the GObject way.

What we now do is g_strdup() all strings that are passed as arguments and then
free the dup'ed strings in the class' finalize() method. Clients must now
manage the strings they pass to these functions themselves.

We will need to bump the ABI version after some more testing.

svn path=/trunk/; revision=4655