Cleanup our Uri handling methods.
* Uri.Fragment is treated differently than Uri.Path (based on when to escape and when to not). Make the API explicit if the fragment needs to be escaped.

* Previous UriToEscapedString (which was converted from a legacy code) did some juggling with the scheme, hostname etc. and then merely appended the path and the fragment. This worked since we always create Uris where the path and the fragment are escaped.

Now, instead of duplicating efforts, we either return the original string used to create the already escaped Uri or we use the AbsoluteUri (which is more correct than uri.ToString()). For our path Uris, we really want the string we used to create the path-uri, so that works as before. For the rest, AbsoluteUri makes sure the Uri is already escaped which is what the earlier method would have done.

svn path=/trunk/; revision=4777