An attempt to address the "too many matches" problem. Replace the oft-unnoticed/multiple-choice scope menu by a short(!) combo box displaying the common search scopes. In the earlier implementation the selected scope menu was used to filter results after a query against all data. In this implementation, use the scope option to search against only the selected type of data. Remove the use of TypeFilter.cs which was asked in the FIXME "it would be nicer to create QueryParts to do the filtering beagled-side".
One problem with this approach is that there are just too many categories (currently ~ 15). Ideally, only the most common categories can be put here and the rest moved to the "Custom search" dialog. Oh right, someone has to create the Custom-search/Advanced-search dialog.

To allow users to add custom scopes, use the environment variables BEAGLE_SCOPE="label1=query-string1;label2=query-string2" and the appropriate query-string is appended to the query when one of these labels are selected in the combobox.

svn path=/trunk/; revision=4792