Fix frequent beagled crashes under mono-2.8 (#641270)
Beagled frequently crashes under mono-2.8 due to uncaught exceptions
in UnixConnectionHandler.WatchCallback() (see Gnome bug #641270 and
http://bugs.gentoo.org/show_bug.cgi?id=353533).
These crashes have two origins.
First, probably due to threading changes in mono, when WatchCallback()
is called, client can be null, so we need to check for it.
Second, presumably due to the exception policy changes in mono-2.8,
client.GetStream().EndRead() legitimately - and fairly frequently -
throws an IOException with Message "Not connected" and with a
null InnerException (instead of a SocketException, as was the case
under previous versions of mono); so we need to catch it.

Signed-off-by: Alexandre Rostovtsev <tetromino@gmail.com>